Added custom header for Prerender requests
It can be undesirable (or impossible) to check for "Prerender" in the `User-Agent` header in some environments. For example, AWS CloudFront can use headers as part of the cache strategy, but caching a different version for every User-Agent string is inefficient and generally unnecessary.

This PR adds a custom header to all requests from Prerender so that they can be identified more easily. This should have no effect on existing behavior.