Fix browser shutdown procedure
The main browser process is now spawn detached, which mean it'll be the
group owner of its children (Chrome will spawn new processes). This way
we can send a SIGINT for the whole group and fix the restart procedure
(before this change prerender could not connect to the browser after a
restart).